@@ -1,78 +1,95 @@
 {
-  "name": "mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap",
-  "type": "library",
-  "description": "Provides viewhelper for Mezzio Navigation to use with Bootstrap",
-  "homepage": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Thomas Müller",
-      "homepage": "https://github.com/mimmi20",
-      "role": "Developer"
+    "name": "mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap",
+    "description": "Provides viewhelper for Mezzio Navigation to use with Bootstrap",
+    "license": "MIT",
+    "type": "library",
+    "authors": [
+        {
+            "name": "Thomas Müller",
+            "homepage": "https://github.com/mimmi20",
+            "role": "Developer"
+        }
+    ],
+    "homepage": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap",
+    "support": {
+        "issues": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap/issues",
+        "source": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap"
+    },
+    "require": {
+        "php": "^8.1.0",
+        "ext-mbstring": "*",
+        "mimmi20/laminasviewrenderer-helper-htmlelement": "^2.0.5",
+        "mimmi20/laminasviewrenderer-helper-partialrenderer": "^2.0.0",
+        "mimmi20/mezzio-navigation": "^3.0.1",
+        "mimmi20/mezzio-navigation-laminasviewrenderer": "^3.0.0",
+        "mimmi20/navigation-helper-acceptpage": "^2.0.0",
+        "mimmi20/navigation-helper-containerparser": "^4.0.0",
+        "mimmi20/navigation-helper-converttopages": "^2.0.0",
+        "mimmi20/navigation-helper-findactive": "^2.0.0",
+        "mimmi20/navigation-helper-findfromproperty": "^2.0.0",
+        "mimmi20/navigation-helper-findroot": "^2.0.0",
+        "mimmi20/navigation-helper-htmlify": "^4.0.0",
+        "psr/log": "^3.0.0",
+        "psr/container": "^1.1.2 || ^2.0.2"
+    },
+    "require-dev": {
+        "ext-ctype": "*",
+        "ext-dom": "*",
+        "ext-intl": "*",
+        "ext-simplexml": "*",
+        "ext-tokenizer": "*",
+        "ext-xml": "*",
+        "ext-xmlreader": "*",
+        "ext-xmlwriter": "*",
+        "infection/infection": "^0.27.9",
+        "laminas/laminas-config": "^3.9.0",
+        "laminas/laminas-escaper": "^2.13.0",
+        "laminas/laminas-i18n": "^2.24.1",
+        "laminas/laminas-servicemanager": "^3.22.1",
+        "mezzio/mezzio-authentication": "^1.9.0",
+        "mezzio/mezzio-router": "^3.7.0",
+        "mimmi20/coding-standard": "^5.2.2",
+        "mimmi20/mezzio-generic-authorization": "^3.0.3",
+        "mimmi20/mezzio-generic-authorization-acl": "^3.0.0",
+        "nikic/php-parser": "^v4.18.0",
+        "phpstan/extension-installer": "^1.3.1",
+        "phpstan/phpstan": "^1.10.50",
+        "phpstan/phpstan-deprecation-rules": "^1.1.4",
+        "phpstan/phpstan-phpunit": "^1.3.15",
+        "phpunit/phpunit": "^10.5.3",
+        "rector/rector": "^0.18.12"
+    },
+    "suggest": {
+        "ext-intl": "required to use the Translator",
+        "laminas/laminas-config": "^3.5.0, to provide page configuration (optional, as arrays and Traversables are also allowed)",
+        "laminas/laminas-i18n": "^2.11.0, required to use the Translator",
+        "laminas/laminas-servicemanager": "^3.6.4, to use the navigation factories",
+        "mimmi20/mezzio-generic-authorization": "^1.0.1, to use authrization roles or privileges"
+    },
+    "autoload": {
+        "psr-4": {
+            "Mimmi20\\Mezzio\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Mimmi20Test\\Mezzio\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\": "tests/"
+        }
+    },
+    "config": {
+        "preferred-install": "dist",
+        "lock": false,
+        "sort-packages": true,
+        "allow-plugins": {
+            "laminas/laminas-dependency-plugin": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "phpstan/extension-installer": true,
+            "infection/extension-installer": true
+        }
+    },
+    "extra": {
+        "laminas": {
+            "config-provider": "Mimmi20\\Mezzio\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\ConfigProvider"
+        }
     }
-  ],
-  "require": {
-    "php": "^7.4.3 || ^8.0.0",
-    "ext-mbstring": "*",
-    "mimmi20/laminasviewrenderer-helper-htmlelement": "^1.0.3",
-    "mimmi20/laminasviewrenderer-helper-partialrenderer": "^1.0.2",
-    "mimmi20/mezzio-navigation": "^2.0.7",
-    "mimmi20/mezzio-navigation-laminasviewrenderer": "^2.0.18",
-    "mimmi20/navigation-helper-acceptpage": "^1.0.0",
-    "mimmi20/navigation-helper-containerparser": "^1.0.0",
-    "mimmi20/navigation-helper-converttopages": "^1.0.0",
-    "mimmi20/navigation-helper-findactive": "^1.0.0",
-    "mimmi20/navigation-helper-findfromproperty": "^1.0.0",
-    "mimmi20/navigation-helper-findroot": "^1.0.0",
-    "mimmi20/navigation-helper-htmlify": "^1.0.0"
-  },
-  "require-dev": {
-    "ext-intl": "*",
-    "laminas/laminas-config": "^3.7.0",
-    "laminas/laminas-dependency-plugin": "^2.2.0",
-    "laminas/laminas-escaper": "^2.9.0",
-    "laminas/laminas-i18n": "^2.12.0",
-    "laminas/laminas-servicemanager": "^3.10.0",
-    "mezzio/mezzio-authentication": "^1.3.0",
-    "mezzio/mezzio-router": "^3.7.0",
-    "mimmi20/coding-standard": "^2.5.1",
-    "mimmi20/mezzio-generic-authorization": "^2.0.5",
-    "mimmi20/mezzio-generic-authorization-acl": "^2.0.5",
-    "phpstan/extension-installer": "^1.1.0",
-    "phpstan/phpstan": "^1.2.0",
-    "phpstan/phpstan-deprecation-rules": "^1.0.0",
-    "phpstan/phpstan-phpunit": "^1.0.0",
-    "phpunit/phpunit": "^9.5.10",
-    "rector/rector": "^0.12.5"
-  },
-  "suggest": {
-    "ext-intl": "required to use the Translator",
-    "laminas/laminas-config": "^3.5.0, to provide page configuration (optional, as arrays and Traversables are also allowed)",
-    "laminas/laminas-i18n": "^2.11.0, required to use the Translator",
-    "laminas/laminas-servicemanager": "^3.6.4, to use the navigation factories",
-    "mimmi20/mezzio-generic-authorization": "^1.0.1, to use authrization roles or privileges"
-  },
-  "config": {
-    "preferred-install": "dist",
-    "sort-packages": true
-  },
-  "extra": {
-    "laminas": {
-      "config-provider": "Mezzio\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\ConfigProvider"
-    }
-  },
-  "autoload": {
-    "psr-4": {
-      "Mezzio\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "MezzioTest\\Navigation\\LaminasView\\View\\Helper\\BootstrapNavigation\\": "tests/"
-    }
-  },
-  "support": {
-    "issues": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap/issues",
-    "source": "https://github.com/mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap"
-  }
 }
