@@ -2,7 +2,7 @@
 /**
  * This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
  *
- * Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -10,32 +10,31 @@
 
 declare(strict_types = 1);
 
-namespace MezzioTest\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare;
+namespace Mimmi20Test\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare;
 
+use Laminas\Config\Exception\InvalidArgumentException;
 use Laminas\Config\Exception\RuntimeException;
-use Laminas\Log\Logger;
 use Laminas\View\Exception\ExceptionInterface;
 use Laminas\View\Helper\EscapeHtml;
 use Laminas\View\Helper\EscapeHtmlAttr;
 use Laminas\View\HelperPluginManager as ViewHelperPluginManager;
-use Mezzio\GenericAuthorization\AuthorizationInterface;
-use Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Menu;
-use Mezzio\Navigation\LaminasView\View\Helper\Navigation\ViewHelperInterface;
-use Mezzio\Navigation\Page\PageFactory;
-use Mezzio\Navigation\Page\PageInterface;
 use Mimmi20\LaminasView\Helper\HtmlElement\Helper\HtmlElementInterface;
 use Mimmi20\LaminasView\Helper\PartialRenderer\Helper\PartialRendererInterface;
+use Mimmi20\Mezzio\GenericAuthorization\AuthorizationInterface;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Menu;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\Navigation\ViewHelperInterface;
+use Mimmi20\Mezzio\Navigation\Page\PageFactory;
+use Mimmi20\Mezzio\Navigation\Page\PageInterface;
 use Mimmi20\NavigationHelper\ContainerParser\ContainerParserInterface;
 use Mimmi20\NavigationHelper\Htmlify\HtmlifyInterface;
 use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Exception;
 use Psr\Container\ContainerExceptionInterface;
-use SebastianBergmann\RecursionContext\InvalidArgumentException;
+use Psr\Log\LoggerInterface;
 
 use function assert;
-use function get_class;
-use function gettype;
-use function is_object;
+use function get_debug_type;
 use function is_string;
 use function rtrim;
 use function sprintf;
@@ -45,32 +44,25 @@ use function trim;
 use const PHP_EOL;
 
 /**
- * Tests Mezzio\Navigation\LaminasView\View\Helper\Navigation\Menu.
- *
- * @group Laminas_View
- * @group Laminas_View_Helper
- * @group Compare
+ * Tests Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\Navigation\Menu.
  */
-final class MenuTest extends AbstractTest
+#[Group('Laminas_View')]
+#[Group('Laminas_View_Helper')]
+#[Group('Compare')]
+final class MenuTest extends AbstractTestCase
 {
-    /**
-     * Class name for view helper to test.
-     */
-    protected string $helperName = Menu::class;
-
     /**
      * View helper
      *
      * @var Menu
      */
-    protected ViewHelperInterface $helper;
+    private ViewHelperInterface $helper;
 
     /**
      * @throws Exception
-     * @throws InvalidArgumentException
      * @throws ExceptionInterface
      * @throws ContainerExceptionInterface
-     * @throws \Laminas\Config\Exception\InvalidArgumentException
+     * @throws InvalidArgumentException
      * @throws RuntimeException
      * @throws \Laminas\Stdlib\Exception\InvalidArgumentException
      */
@@ -87,8 +79,8 @@ final class MenuTest extends AbstractTest
             sprintf(
                 '$renderer should be an Instance of %s, but was %s',
                 PartialRendererInterface::class,
-                is_object($renderer) ? get_class($renderer) : gettype($renderer)
-            )
+                get_debug_type($renderer),
+            ),
         );
 
         $escapeHtmlAttr = $plugin->get(EscapeHtmlAttr::class);
@@ -97,15 +89,15 @@ final class MenuTest extends AbstractTest
             sprintf(
                 '$escapeHtmlAttr should be an Instance of %s, but was %s',
                 EscapeHtmlAttr::class,
-                is_object($escapeHtmlAttr) ? get_class($escapeHtmlAttr) : gettype($escapeHtmlAttr)
-            )
+                get_debug_type($escapeHtmlAttr),
+            ),
         );
 
-        $logger          = $this->serviceManager->get(Logger::class);
+        $logger          = $this->serviceManager->get(LoggerInterface::class);
         $htmlify         = $this->serviceManager->get(HtmlifyInterface::class);
         $containerParser = $this->serviceManager->get(ContainerParserInterface::class);
 
-        assert($logger instanceof Logger);
+        assert($logger instanceof LoggerInterface);
         assert($htmlify instanceof HtmlifyInterface);
         assert($containerParser instanceof ContainerParserInterface);
 
@@ -115,8 +107,8 @@ final class MenuTest extends AbstractTest
             sprintf(
                 '$escapeHelper should be an Instance of %s, but was %s',
                 EscapeHtml::class,
-                is_object($escapeHtml) ? get_class($escapeHtml) : gettype($escapeHtml)
-            )
+                get_debug_type($escapeHtml),
+            ),
         );
 
         $htmlElement = $this->serviceManager->get(HtmlElementInterface::class);
@@ -132,7 +124,7 @@ final class MenuTest extends AbstractTest
             $renderer,
             $escapeHtml,
             $htmlElement,
-            null
+            null,
         );
 
         // set nav1 in helper as default
@@ -142,6 +134,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testCanRenderMenuFromServiceAlias(): void
     {
@@ -172,7 +165,7 @@ final class MenuTest extends AbstractTest
      */
     public function testHelperEntryPointWithoutAnyParams(): void
     {
-        $returned = $this->helper->__invoke();
+        $returned = ($this->helper)();
         self::assertSame($this->helper, $returned);
         self::assertSame($this->nav1, $returned->getContainer());
     }
@@ -183,7 +176,7 @@ final class MenuTest extends AbstractTest
      */
     public function testHelperEntryPointWithContainerParam(): void
     {
-        $returned = $this->helper->__invoke($this->nav2);
+        $returned = ($this->helper)($this->nav2);
         self::assertSame($this->helper, $returned);
         self::assertSame($this->nav2, $returned->getContainer());
     }
@@ -201,6 +194,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetIndentAndOverrideInRenderMenu(): void
     {
@@ -223,6 +217,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderSuppliedContainerWithoutInterfering(): void
     {
@@ -247,6 +242,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testUseAclRoleAsString(): void
     {
@@ -266,6 +262,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testFilterOutPagesBasedOnAcl(): void
     {
@@ -285,6 +282,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testDisablingAcl(): void
     {
@@ -304,6 +302,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testDisablingAclWhenUsingUl(): void
     {
@@ -323,6 +322,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testDisablingAclWhenUsingOl(): void
     {
@@ -342,6 +342,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testDisablingAclWhenUsingButton(): void
     {
@@ -353,7 +354,10 @@ final class MenuTest extends AbstractTest
         $this->helper->setUseAuthorization(false);
 
         $expected = $this->getExpected('menu/default1_button.html');
-        $actual   = $this->helper->renderMenu(null, ['style' => Menu::STYLE_UL, 'sublink' => Menu::STYLE_SUBLINK_BUTTON]);
+        $actual   = $this->helper->renderMenu(
+            null,
+            ['style' => Menu::STYLE_UL, 'sublink' => Menu::STYLE_SUBLINK_BUTTON],
+        );
 
         self::assertSame($expected, trim($actual));
     }
@@ -361,6 +365,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testDisablingAclWhenUsingDetails(): void
     {
@@ -372,7 +377,10 @@ final class MenuTest extends AbstractTest
         $this->helper->setUseAuthorization(false);
 
         $expected = $this->getExpected('menu/default1_details.html');
-        $actual   = $this->helper->renderMenu(null, ['style' => Menu::STYLE_UL, 'sublink' => Menu::STYLE_SUBLINK_DETAILS]);
+        $actual   = $this->helper->renderMenu(
+            null,
+            ['style' => Menu::STYLE_UL, 'sublink' => Menu::STYLE_SUBLINK_DETAILS],
+        );
 
         self::assertSame($expected, trim($actual));
     }
@@ -381,6 +389,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetUlCssClass(): void
     {
@@ -396,6 +405,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetLiActiveCssClass(): void
     {
@@ -410,6 +420,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionEscapeLabelsAsTrue(): void
     {
@@ -420,7 +431,7 @@ final class MenuTest extends AbstractTest
             [
                 'label' => 'Badges <span class="badge">1</span>',
                 'uri' => 'badges',
-            ]
+            ],
         );
 
         $nav2->addPage($page);
@@ -434,6 +445,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionEscapeLabelsAsFalse(): void
     {
@@ -444,7 +456,7 @@ final class MenuTest extends AbstractTest
             [
                 'label' => 'Badges <span class="badge">1</span>',
                 'uri' => 'badges',
-            ]
+            ],
         );
 
         $nav2->addPage($page);
@@ -459,6 +471,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \InvalidArgumentException
      * @throws \Laminas\View\Exception\RuntimeException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderingPartial(): void
     {
@@ -474,6 +487,7 @@ final class MenuTest extends AbstractTest
      * @throws Exception
      * @throws \Laminas\View\Exception\RuntimeException
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderingPartialBySpecifyingAnArrayAsPartial(): void
     {
@@ -503,6 +517,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws AssertionFailedError
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderingPartialShouldFailOnInvalidPartialArray(): void
     {
@@ -511,13 +526,14 @@ final class MenuTest extends AbstractTest
         try {
             $this->helper->render();
             self::fail('invalid $partial should throw Laminas\View\Exception\InvalidArgumentException');
-        } catch (ExceptionInterface $e) {
+        } catch (ExceptionInterface) {
         }
     }
 
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetMaxDepth(): void
     {
@@ -532,6 +548,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetMinDepth(): void
     {
@@ -546,6 +563,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetBothDepts(): void
     {
@@ -560,6 +578,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetOnlyActiveBranch(): void
     {
@@ -574,6 +593,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetRenderParents(): void
     {
@@ -588,6 +608,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testSetOnlyActiveBranchAndMinDepth(): void
     {
@@ -602,6 +623,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOnlyActiveBranchAndMaxDepth(): void
     {
@@ -616,6 +638,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOnlyActiveBranchAndBothDepthsSpecified(): void
     {
@@ -630,6 +653,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOnlyActiveBranchNoParentsAndBothDepthsSpecified(): void
     {
@@ -647,6 +671,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOnlyActiveBranchNoParentsActiveOneBelowMinDepth(): void
     {
@@ -666,6 +691,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderSubMenuShouldOverrideOptions(): void
     {
@@ -683,6 +709,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionMaxDepth(): void
     {
@@ -697,6 +724,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionMinDepth(): void
     {
@@ -711,6 +739,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionBothDepts(): void
     {
@@ -728,6 +757,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranch(): void
     {
@@ -742,6 +772,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranchNoParents(): void
     {
@@ -759,6 +790,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranchAndMinDepth(): void
     {
@@ -776,6 +808,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranchAndMaxDepth(): void
     {
@@ -793,6 +826,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranchAndBothDepthsSpecified(): void
     {
@@ -811,6 +845,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testOptionOnlyActiveBranchNoParentsAndBothDepthsSpecified(): void
     {
@@ -830,6 +865,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderingWithoutPageClassToLi(): void
     {
@@ -839,7 +875,7 @@ final class MenuTest extends AbstractTest
                 'label' => 'Class test',
                 'uri' => 'test',
                 'class' => 'foobar',
-            ]
+            ],
         );
 
         $nav2->addPage($page);
@@ -853,6 +889,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderingWithPageClassToLi(): void
     {
@@ -864,7 +901,7 @@ final class MenuTest extends AbstractTest
                 'label' => 'Class test',
                 'uri' => 'test',
                 'class' => 'foobar',
-            ]
+            ],
         );
         $nav2->addPage($page);
 
@@ -877,6 +914,7 @@ final class MenuTest extends AbstractTest
     /**
      * @throws Exception
      * @throws \InvalidArgumentException
+     * @throws \Laminas\I18n\Exception\RuntimeException
      */
     public function testRenderDeepestMenuWithPageClassToLi(): void
     {
@@ -894,8 +932,8 @@ final class MenuTest extends AbstractTest
             sprintf(
                 '$page should be an Instance of %s, but was %s',
                 PageInterface::class,
-                get_class($page)
-            )
+                $page::class,
+            ),
         );
 
         self::assertInstanceOf(PageInterface::class, $page);
@@ -909,12 +947,19 @@ final class MenuTest extends AbstractTest
 
     /**
      * Returns the contens of the expected $file, normalizes newlines.
+     *
+     * @throws Exception
      */
     protected function getExpected(string $file): string
     {
-        return str_replace(["\r\n", "\n", "\r", '##lb##'], ['##lb##', '##lb##', '##lb##', PHP_EOL], parent::getExpected($file));
+        return str_replace(
+            ["\r\n", "\n", "\r", '##lb##'],
+            ['##lb##', '##lb##', '##lb##', PHP_EOL],
+            parent::getExpected($file),
+        );
     }
 
+    /** @throws void */
     private function setActive(string $label): void
     {
         $container = $this->helper->getContainer();
