@@ -2,7 +2,7 @@
 /**
  * This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
  *
- * Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,7 @@ declare(strict_types = 1);
 $header = <<<'EOF'
     This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
 
-    Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+    Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
 
     For the full copyright and license information, please view the LICENSE
     file that was distributed with this source code.
@@ -24,6 +24,7 @@ $finder = PhpCsFixer\Finder::create()
     ->name('*.php')
     ->in(__DIR__ . '/src')
     ->in(__DIR__ . '/tests')
+    ->append([__DIR__ . '/rector.php'])
     ->append([__FILE__]);
 
 $rules = require 'vendor/mimmi20/coding-standard/src/php-cs-fixer.config.php';
@@ -42,8 +43,8 @@ return $config
                     'location' => 'after_open',
                     'separate' => 'bottom',
                 ],
-            ]
-        )
+            ],
+        ),
     )
     ->setUsingCache(true)
     ->setFinder($finder);
