@@ -2,7 +2,7 @@
 /**
  * This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
  *
- * Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -10,22 +10,23 @@
 
 declare(strict_types = 1);
 
-namespace MezzioTest\Navigation\LaminasView\View\Helper\BootstrapNavigation;
+namespace Mimmi20Test\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation;
 
 use Interop\Container\ContainerInterface;
 use Laminas\I18n\View\Helper\Translate;
-use Laminas\Log\Logger;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 use Laminas\View\Helper\EscapeHtml;
+use Laminas\View\Helper\HelperInterface;
 use Laminas\View\HelperPluginManager as ViewHelperPluginManager;
-use Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Breadcrumbs;
-use Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\BreadcrumbsFactory;
 use Mimmi20\LaminasView\Helper\PartialRenderer\Helper\PartialRendererInterface;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Breadcrumbs;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\BreadcrumbsFactory;
 use Mimmi20\NavigationHelper\ContainerParser\ContainerParserInterface;
 use Mimmi20\NavigationHelper\Htmlify\HtmlifyInterface;
 use PHPUnit\Framework\Exception;
 use PHPUnit\Framework\TestCase;
-use SebastianBergmann\RecursionContext\InvalidArgumentException;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Log\LoggerInterface;
 
 use function assert;
 
@@ -33,6 +34,7 @@ final class BreadcrumbsFactoryTest extends TestCase
 {
     private BreadcrumbsFactory $factory;
 
+    /** @throws void */
     protected function setUp(): void
     {
         $this->factory = new BreadcrumbsFactory();
@@ -40,23 +42,23 @@ final class BreadcrumbsFactoryTest extends TestCase
 
     /**
      * @throws Exception
-     * @throws InvalidArgumentException
+     * @throws ContainerExceptionInterface
      */
     public function testInvocationWithTranslator(): void
     {
-        $logger = $this->getMockBuilder(Logger::class)
+        $logger = $this->getMockBuilder(LoggerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
         $logger->expects(self::never())
-            ->method('emerg');
+            ->method('emergency');
         $logger->expects(self::never())
             ->method('alert');
         $logger->expects(self::never())
-            ->method('crit');
+            ->method('critical');
         $logger->expects(self::never())
-            ->method('err');
+            ->method('error');
         $logger->expects(self::never())
-            ->method('warn');
+            ->method('warning');
         $logger->expects(self::never())
             ->method('notice');
         $logger->expects(self::never())
@@ -77,18 +79,50 @@ final class BreadcrumbsFactoryTest extends TestCase
             ->method('has')
             ->with(Translate::class)
             ->willReturn(true);
-        $viewHelperPluginManager->expects(self::exactly(2))
+        $matcher = self::exactly(2);
+        $viewHelperPluginManager->expects($matcher)
             ->method('get')
-            ->withConsecutive([Translate::class], [EscapeHtml::class])
-            ->willReturnOnConsecutiveCalls($translatePlugin, $escapePlugin);
+            ->willReturnCallback(
+                static function (string $name, array | null $options = null) use ($matcher, $translatePlugin, $escapePlugin): HelperInterface {
+                    match ($matcher->numberOfInvocations()) {
+                        1 => self::assertSame(Translate::class, $name),
+                        default => self::assertSame(EscapeHtml::class, $name),
+                    };
+
+                    self::assertNull($options);
+
+                    return match ($matcher->numberOfInvocations()) {
+                        1 => $translatePlugin,
+                        default => $escapePlugin,
+                    };
+                },
+            );
 
         $container = $this->getMockBuilder(ServiceLocatorInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $container->expects(self::exactly(5))
+        $matcher   = self::exactly(5);
+        $container->expects($matcher)
             ->method('get')
-            ->withConsecutive([ViewHelperPluginManager::class], [Logger::class], [HtmlifyInterface::class], [ContainerParserInterface::class], [PartialRendererInterface::class])
-            ->willReturnOnConsecutiveCalls($viewHelperPluginManager, $logger, $htmlify, $containerParser, $renderer);
+            ->willReturnCallback(
+                static function (string $id) use ($matcher, $viewHelperPluginManager, $logger, $htmlify, $containerParser, $renderer): mixed {
+                    match ($matcher->numberOfInvocations()) {
+                        1 => self::assertSame(ViewHelperPluginManager::class, $id),
+                        2 => self::assertSame(LoggerInterface::class, $id),
+                        3 => self::assertSame(HtmlifyInterface::class, $id),
+                        4 => self::assertSame(ContainerParserInterface::class, $id),
+                        default => self::assertSame(PartialRendererInterface::class, $id),
+                    };
+
+                    return match ($matcher->numberOfInvocations()) {
+                        1 => $viewHelperPluginManager,
+                        2 => $logger,
+                        3 => $htmlify,
+                        4 => $containerParser,
+                        default => $renderer,
+                    };
+                },
+            );
 
         assert($container instanceof ContainerInterface);
         $helper = ($this->factory)($container);
@@ -98,23 +132,23 @@ final class BreadcrumbsFactoryTest extends TestCase
 
     /**
      * @throws Exception
-     * @throws InvalidArgumentException
+     * @throws ContainerExceptionInterface
      */
     public function testInvocationWithoutTranslator(): void
     {
-        $logger = $this->getMockBuilder(Logger::class)
+        $logger = $this->getMockBuilder(LoggerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
         $logger->expects(self::never())
-            ->method('emerg');
+            ->method('emergency');
         $logger->expects(self::never())
             ->method('alert');
         $logger->expects(self::never())
-            ->method('crit');
+            ->method('critical');
         $logger->expects(self::never())
-            ->method('err');
+            ->method('error');
         $logger->expects(self::never())
-            ->method('warn');
+            ->method('warning');
         $logger->expects(self::never())
             ->method('notice');
         $logger->expects(self::never())
@@ -142,10 +176,28 @@ final class BreadcrumbsFactoryTest extends TestCase
         $container = $this->getMockBuilder(ServiceLocatorInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $container->expects(self::exactly(5))
+        $matcher   = self::exactly(5);
+        $container->expects($matcher)
             ->method('get')
-            ->withConsecutive([ViewHelperPluginManager::class], [Logger::class], [HtmlifyInterface::class], [ContainerParserInterface::class], [PartialRendererInterface::class])
-            ->willReturnOnConsecutiveCalls($viewHelperPluginManager, $logger, $htmlify, $containerParser, $renderer);
+            ->willReturnCallback(
+                static function (string $id) use ($matcher, $viewHelperPluginManager, $logger, $htmlify, $containerParser, $renderer): mixed {
+                    match ($matcher->numberOfInvocations()) {
+                        1 => self::assertSame(ViewHelperPluginManager::class, $id),
+                        2 => self::assertSame(LoggerInterface::class, $id),
+                        3 => self::assertSame(HtmlifyInterface::class, $id),
+                        4 => self::assertSame(ContainerParserInterface::class, $id),
+                        default => self::assertSame(PartialRendererInterface::class, $id),
+                    };
+
+                    return match ($matcher->numberOfInvocations()) {
+                        1 => $viewHelperPluginManager,
+                        2 => $logger,
+                        3 => $htmlify,
+                        4 => $containerParser,
+                        default => $renderer,
+                    };
+                },
+            );
 
         assert($container instanceof ContainerInterface);
         $helper = ($this->factory)($container);
