@@ -2,7 +2,7 @@
 /**
  * This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
  *
- * Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -10,13 +10,13 @@
 
 declare(strict_types = 1);
 
-namespace MezzioTest\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare;
+namespace Mimmi20Test\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare;
 
 use Laminas\Config\Config;
+use Laminas\Config\Exception\InvalidArgumentException;
 use Laminas\Config\Exception\RuntimeException;
 use Laminas\Config\Factory as ConfigFactory;
 use Laminas\I18n\Translator\Translator;
-use Laminas\Log\Logger;
 use Laminas\Permissions\Acl\Acl;
 use Laminas\Permissions\Acl\Resource\GenericResource;
 use Laminas\Permissions\Acl\Role\GenericRole;
@@ -24,7 +24,6 @@ use Laminas\ServiceManager\Exception\ContainerModificationsNotAllowedException;
 use Laminas\ServiceManager\Factory\InvokableFactory;
 use Laminas\ServiceManager\ServiceManager;
 use Laminas\View\HelperPluginManager as ViewHelperPluginManager;
-use Mezzio\GenericAuthorization\Acl\LaminasAcl;
 use Mezzio\Helper\ServerUrlHelper as BaseServerUrlHelper;
 use Mezzio\Helper\UrlHelper as BaseUrlHelper;
 use Mezzio\LaminasView\HelperPluginManagerFactory;
@@ -32,23 +31,23 @@ use Mezzio\LaminasView\LaminasViewRenderer;
 use Mezzio\LaminasView\LaminasViewRendererFactory;
 use Mezzio\LaminasView\ServerUrlHelper;
 use Mezzio\LaminasView\UrlHelper;
-use Mezzio\Navigation\Config\NavigationConfig;
-use Mezzio\Navigation\Config\NavigationConfigInterface;
-use Mezzio\Navigation\LaminasView\View\Helper\Navigation\ViewHelperInterface;
-use Mezzio\Navigation\LaminasView\View\Helper\NavigationFactory;
-use Mezzio\Navigation\LaminasView\View\Helper\ServerUrlHelperFactory;
-use Mezzio\Navigation\LaminasView\View\Helper\UrlHelperFactory;
-use Mezzio\Navigation\Navigation;
-use Mezzio\Navigation\Page\PageFactory;
-use Mezzio\Navigation\Page\PageFactoryInterface;
-use Mezzio\Navigation\Service\ConstructedNavigationFactory;
-use Mezzio\Navigation\Service\DefaultNavigationFactory;
 use Mezzio\Router\Route;
 use Mezzio\Router\RouteResult;
 use Mimmi20\LaminasView\Helper\HtmlElement\Helper\HtmlElementFactory;
 use Mimmi20\LaminasView\Helper\HtmlElement\Helper\HtmlElementInterface;
 use Mimmi20\LaminasView\Helper\PartialRenderer\Helper\PartialRendererFactory;
 use Mimmi20\LaminasView\Helper\PartialRenderer\Helper\PartialRendererInterface;
+use Mimmi20\Mezzio\GenericAuthorization\Acl\LaminasAcl;
+use Mimmi20\Mezzio\Navigation\Config\NavigationConfig;
+use Mimmi20\Mezzio\Navigation\Config\NavigationConfigInterface;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\NavigationFactory;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\ServerUrlHelperFactory;
+use Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\UrlHelperFactory;
+use Mimmi20\Mezzio\Navigation\Navigation;
+use Mimmi20\Mezzio\Navigation\Page\PageFactory;
+use Mimmi20\Mezzio\Navigation\Page\PageFactoryInterface;
+use Mimmi20\Mezzio\Navigation\Service\ConstructedNavigationFactory;
+use Mimmi20\Mezzio\Navigation\Service\DefaultNavigationFactory;
 use Mimmi20\NavigationHelper\Accept\AcceptHelperFactory;
 use Mimmi20\NavigationHelper\Accept\AcceptHelperInterface;
 use Mimmi20\NavigationHelper\ContainerParser\ContainerParserFactory;
@@ -67,19 +66,17 @@ use PHPUnit\Framework\Exception;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Http\Server\MiddlewareInterface;
-use SebastianBergmann\RecursionContext\InvalidArgumentException;
+use Psr\Log\LoggerInterface;
 
 use function assert;
 use function file_get_contents;
-use function get_class;
-use function gettype;
-use function is_object;
+use function get_debug_type;
 use function sprintf;
 
 /**
  * Base class for navigation view helper tests
  */
-abstract class AbstractTest extends TestCase
+abstract class AbstractTestCase extends TestCase
 {
     protected ServiceManager $serviceManager;
 
@@ -87,16 +84,6 @@ abstract class AbstractTest extends TestCase
      * Path to files needed for test
      */
     protected string $files;
-
-    /**
-     * Class name for view helper to test
-     */
-    protected string $helperName;
-
-    /**
-     * View helper
-     */
-    protected ViewHelperInterface $helper;
 
     /**
      * The first container in the config file (files/navigation.xml)
@@ -117,9 +104,8 @@ abstract class AbstractTest extends TestCase
      * Prepares the environment before running a test
      *
      * @throws Exception
-     * @throws InvalidArgumentException
      * @throws ContainerExceptionInterface
-     * @throws \Laminas\Config\Exception\InvalidArgumentException
+     * @throws InvalidArgumentException
      * @throws RuntimeException
      */
     protected function setUp(): void
@@ -146,7 +132,7 @@ abstract class AbstractTest extends TestCase
             function () use ($config): NavigationConfig {
                 $route = new Route(
                     '/test.html',
-                    $this->createMock(MiddlewareInterface::class)
+                    $this->createMock(MiddlewareInterface::class),
                 );
 
                 $pages            = $config->toArray();
@@ -159,11 +145,11 @@ abstract class AbstractTest extends TestCase
                     [
                         'route' => 'post',
                         'id' => '1337',
-                    ]
+                    ],
                 ));
 
                 return $navConfig;
-            }
+            },
         );
         $sm->setFactory(PageFactory::class, InvokableFactory::class);
         $sm->setAlias(PageFactoryInterface::class, PageFactory::class);
@@ -185,7 +171,7 @@ abstract class AbstractTest extends TestCase
                 ],
                 'view_helpers' => [
                     'aliases' => [
-                        'navigation' => \Mezzio\Navigation\LaminasView\View\Helper\Navigation::class,
+                        'navigation' => \Mimmi20\Mezzio\Navigation\LaminasView\View\Helper\Navigation::class,
                         'Navigation' => Navigation::class,
                         BaseServerUrlHelper::class => ServerUrlHelper::class,
                         'serverurl' => ServerUrlHelper::class,
@@ -210,25 +196,25 @@ abstract class AbstractTest extends TestCase
                         'test::bc-with-partials' => __DIR__ . '/_files/mvc/views/bc_with_partial_params.phtml',
                     ],
                 ],
-            ]
+            ],
         );
         $sm->setFactory(ViewHelperPluginManager::class, HelperPluginManagerFactory::class);
         $sm->setFactory(LaminasViewRenderer::class, LaminasViewRendererFactory::class);
         $sm->setFactory(BaseServerUrlHelper::class, InvokableFactory::class);
 
-        $logger = $this->getMockBuilder(Logger::class)
+        $logger = $this->getMockBuilder(LoggerInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
         $logger->expects(static::never())
-            ->method('emerg');
+            ->method('emergency');
         $logger->expects(static::never())
             ->method('alert');
         $logger->expects(static::never())
-            ->method('crit');
+            ->method('critical');
         $logger->expects(static::never())
-            ->method('err');
+            ->method('error');
         $logger->expects(static::never())
-            ->method('warn');
+            ->method('warning');
         $logger->expects(static::never())
             ->method('notice');
         $logger->expects(static::never())
@@ -236,7 +222,7 @@ abstract class AbstractTest extends TestCase
         $logger->expects(static::never())
             ->method('debug');
 
-        $sm->setService(Logger::class, $logger);
+        $sm->setService(LoggerInterface::class, $logger);
 
         // setup containers from config
         $nav1 = $sm->get('nav_test1');
@@ -248,8 +234,8 @@ abstract class AbstractTest extends TestCase
             sprintf(
                 '$nav1 should be an Instance of %s, but was %s',
                 Navigation::class,
-                is_object($nav1) ? get_class($nav1) : gettype($nav1)
-            )
+                get_debug_type($nav1),
+            ),
         );
         assert($nav2 instanceof Navigation);
         assert($nav3 instanceof Navigation);
@@ -268,13 +254,15 @@ abstract class AbstractTest extends TestCase
      * Returns the contens of the expected $file
      *
      * @throws Exception
-     * @throws InvalidArgumentException
      */
     protected function getExpected(string $file): string
     {
         $content = file_get_contents($this->files . '/expected/' . $file);
 
-        static::assertIsString($content, sprintf('could not load file %s', $this->files . '/expected/' . $file));
+        static::assertIsString(
+            $content,
+            sprintf('could not load file %s', $this->files . '/expected/' . $file),
+        );
 
         return $content;
     }
@@ -329,7 +317,7 @@ abstract class AbstractTest extends TestCase
 
         $translator = new Translator();
         $translator->getPluginManager()->setService('default', $loader);
-        $translator->addTranslationFile('default', null);
+        $translator->addTranslationFile('default', '');
 
         return $translator;
     }
@@ -364,8 +352,8 @@ abstract class AbstractTest extends TestCase
         $translator = new Translator();
         $translator->getPluginManager()->setService('default1', $loader1);
         $translator->getPluginManager()->setService('default2', $loader2);
-        $translator->addTranslationFile('default1', null, 'LaminasTest_1');
-        $translator->addTranslationFile('default2', null, 'LaminasTest_2');
+        $translator->addTranslationFile('default1', '', 'LaminasTest_1');
+        $translator->addTranslationFile('default2', 'null', 'LaminasTest_2');
 
         return $translator;
     }
