@@ -2,7 +2,7 @@
 /**
  * This file is part of the mimmi20/mezzio-navigation-laminasviewrenderer-bootstrap package.
  *
- * Copyright (c) 2021, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2021-2023, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -10,19 +10,17 @@
 
 declare(strict_types = 1);
 
-/**
- * @see       https://github.com/laminas/laminas-view for the canonical source repository
- */
+/** @see       https://github.com/laminas/laminas-view for the canonical source repository */
 
-namespace MezzioTest\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare\TestAsset;
+namespace Mimmi20Test\Mezzio\Navigation\LaminasView\View\Helper\BootstrapNavigation\Compare\TestAsset;
 
 use Laminas\I18n\Translator;
-use Laminas\I18n\Translator\TextDomain;
 
+/** phpcs:disable SlevomatCodingStandard.Classes.ForbiddenPublicProperty.ForbiddenPublicProperty */
 final class ArrayTranslator implements Translator\Loader\FileLoaderInterface
 {
-    /** @var array<string> */
-    public array $translations;
+    /** @var array<string, string> */
+    public array $translations = [];
 
     /**
      * Load translations from a file.
@@ -30,10 +28,14 @@ final class ArrayTranslator implements Translator\Loader\FileLoaderInterface
      * @param string $locale
      * @param string $filename
      *
+     * @return Translator\TextDomain<string, string>
+     *
+     * @throws void
+     *
      * @phpcsSuppress SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingNativeTypeHint
      * @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter
      */
-    public function load($filename, $locale): ?TextDomain
+    public function load($filename, $locale): Translator\TextDomain
     {
         return new Translator\TextDomain($this->translations);
     }
